Refactor Wagtail API key handling and service environment setup

+++ init.sh
@@ -50,21 +50,24 @@ do
         echo
     fi
 
-    if [[ -f "services/$service/.env.example" ]]
+    if [[ ! -f "services/$service/.env" ]]
     then
-        echo "Copying example .env for $service..."
-        cp "services/$service/.env.example" "services/$service/.env"
-        echo
-    elif [[ -f "services/$service/.example.env" ]]
-    then
-        echo "Copying example .env for $service..."
-        cp "services/$service/.example.env" "services/$service/.env"
-        echo
-    # elif [[ ! -f "services/$service/.env" ]]
-    # then
-    #     echo "Creating blank .env for $service..."
-    #     touch "services/$service/.env"
-    #     echo
+        if [[ -f "services/$service/.env.example" ]]
+        then
+            echo "Copying example .env for $service..."
+            cp "services/$service/.env.example" "services/$service/.env"
+            echo
+        elif [[ -f "services/$service/.example.env" ]]
+        then
+            echo "Copying example .env for $service..."
+            cp "services/$service/.example.env" "services/$service/.env"
+            echo
+        # elif [[ ! -f "services/$service/.env" ]]
+        # then
+        #     echo "Creating blank .env for $service..."
+        #     touch "services/$service/.env"
+        #     echo
+        fi
     fi
 done
